Split repository.ts: extract membership + identity linking (9/N)
Ninth slice of the repository.ts carve-up (audit L14). Moves two adjacent
sections into repository/members.ts: the three-tier RBAC rows in
`community_users` plus the auto-enrol path, and cross-platform identity linking
(one person holding both a Discord and a WhatsApp id).

The two sections move TOGETHER deliberately: the private
`dissolveGroupIfUnderTwo` helper is defined in the membership half and called
from the linking half, so splitting them would have forced a shared-internals
export for a helper that is nobody else's business. It travels verbatim with
its callers.

Three couplings that the call-pattern screen missed but typecheck caught:

- members.ts calls `recordAdminAction`, which moved to repository/adminAudit.ts
  in #812 — so members.ts imports it from there. This is the first
  domain-module-to-domain-module import in the campaign; it is not a cycle
  (adminAudit imports nothing from members).
- `AUTO_ENROLL_ACTOR` is defined in the membership half but also read by
  adminActivitySummary, which stays — so repository.ts imports it explicitly
  alongside `resolveLinkedIdentities` (also still used by three functions that
  stay). `export *` does not bind names locally.
- `PoolClient` and the previous `recordAdminAction` import became unused in
  repository.ts and were dropped.

All 17 moved symbols verified byte-identical against origin/main, including the
private helper. No conversation-scoped admin read is involved, so the 🔒
SQL-scoping invariant is untouched — but note this module now owns the
`community_users` rows the RBAC layer treats as its source of truth, so a
future change here is security-relevant even though this move is not.

repository.ts is now 4,681 lines, down from 7,137 at the start of the campaign.
